Crate google-storage1

This documentation was generated from storage crate version 0.0.1. The original source code can be found on github.

Features

Handle the following Resources with ease from the central hub ...

bucket access controls

delete, get, insert, list, patch and update

buckets

delete, get, insert, list, patch and update

channels

stop

default object access controls

delete, get, insert, list, patch and update

object access controls

delete, get, insert, list, patch and update

objects

compose, copy, delete, get, insert, list, patch, update and watch all

Upload supported by ...

insert objects

Download supported by ...

get objects

update objects

insert objects

compose objects

copy objects

Subscription supported by ...

watch all objects

list objects

Everything else about the storage v1 API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.object_access_controls().get(...).doit()
let r = hub.default_object_access_controls().patch(...).doit()
let r = hub.object_access_controls().update(...).doit()
let r = hub.object_access_controls().list(...).doit()
let r = hub.object_access_controls().patch(...).doit()
let r = hub.default_object_access_controls().update(...).doit()
let r = hub.default_object_access_controls().insert(...).doit()
let r = hub.object_access_controls().insert(...).doit()
let r = hub.object_access_controls().delete(...).doit()
let r = hub.default_object_access_controls().get(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-storage1 = "0.0.1"

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate serde;
extern crate "google-storage1" as storage1;
use storage1::ObjectAccessControl;
use storage1::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use storage1::Storage;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Storage::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req: ObjectAccessControl = Default::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.object_access_controls().update(&req, "bucket", "object", "entity")
             .generation("accusam")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::MissingAPIKey => println!("Missing API Key"),
    Result::MissingToken => println!("Missing Token"),
    Result::Failure(_) => println!("General Failure (Response doesn't print)"),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::JsonDecodeError(err) => println!("Json failed to decode: {:?}", err),
    Result::Success(_) => println!("Success (value doesn't print)"),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Uploads and Downlods

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

Bucket
A bucket.

BucketAccessControl
An access-control entry.

BucketAccessControlDeleteCallBuilder
Permanently deletes the ACL entry for the specified entity on the specified bucket.

BucketAccessControlGetCallBuilder
Returns the ACL entry for the specified entity on the specified bucket.

BucketAccessControlInsertCallBuilder
Creates a new ACL entry on the specified bucket.

BucketAccessControlListCallBuilder
Retrieves ACL entries on the specified bucket.

BucketAccessControlMethodsBuilder
A builder providing access to all methods supported on bucketAccessControl resources. It is not used directly, but through the Storage hub.

BucketAccessControlPatchCallBuilder
Updates an ACL entry on the specified bucket. This method supports patch semantics.

BucketAccessControlProjectTeam
The project team associated with the entity, if any.

BucketAccessControlUpdateCallBuilder
Updates an ACL entry on the specified bucket.

BucketAccessControls
An access-control list.

BucketCors
The bucket's Cross-Origin Resource Sharing (CORS) configuration.

BucketDeleteCallBuilder
Permanently deletes an empty bucket.

BucketGetCallBuilder
Returns metadata for the specified bucket.

BucketInsertCallBuilder
Creates a new bucket.

BucketLifecycle
The bucket's lifecycle configuration. See lifecycle management for more information.

BucketLifecycleRule
A lifecycle management rule, which is made of an action to take and the condition(s) under which the action will be taken.

BucketLifecycleRuleAction
The action to take.

BucketLifecycleRuleCondition
The condition(s) under which the action will be taken.

BucketListCallBuilder
Retrieves a list of buckets for a given project.

BucketLogging
The bucket's logging configuration, which defines the destination bucket and optional name prefix for the current bucket's logs.

BucketMethodsBuilder
A builder providing access to all methods supported on bucket resources. It is not used directly, but through the Storage hub.

BucketOwner
The owner of the bucket. This is always the project team's owner group.

BucketPatchCallBuilder
Updates a bucket. This method supports patch semantics.

BucketUpdateCallBuilder
Updates a bucket.

BucketVersioning
The bucket's versioning configuration.

BucketWebsite
The bucket's website configuration.

Buckets
A list of buckets.

Channel
An notification channel used to watch for resource changes.

ChannelMethodsBuilder
A builder providing access to all methods supported on channel resources. It is not used directly, but through the Storage hub.

ChannelStopCallBuilder
Stop watching resources through this channel

ComposeRequest
A Compose request.

ComposeRequestSourceObjects
The list of source objects that will be concatenated into a single object.

ComposeRequestSourceObjectsObjectPreconditions
Conditions that must be met for this operation to execute.

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

DefaultObjectAccessControlDeleteCallBuilder
Permanently deletes the default object ACL entry for the specified entity on the specified bucket.

DefaultObjectAccessControlGetCallBuilder
Returns the default object ACL entry for the specified entity on the specified bucket.

DefaultObjectAccessControlInsertCallBuilder
Creates a new default object ACL entry on the specified bucket.

DefaultObjectAccessControlListCallBuilder
Retrieves default object ACL entries on the specified bucket.

DefaultObjectAccessControlMethodsBuilder
A builder providing access to all methods supported on defaultObjectAccessControl resources. It is not used directly, but through the Storage hub.

DefaultObjectAccessControlPatchCallBuilder
Updates a default object ACL entry on the specified bucket. This method supports patch semantics.

DefaultObjectAccessControlUpdateCallBuilder
Updates a default object ACL entry on the specified bucket.

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

Object
An object.

ObjectAccessControl
An access-control entry.

ObjectAccessControlDeleteCallBuilder
Permanently deletes the ACL entry for the specified entity on the specified object.

ObjectAccessControlGetCallBuilder
Returns the ACL entry for the specified entity on the specified object.

ObjectAccessControlInsertCallBuilder
Creates a new ACL entry on the specified object.

ObjectAccessControlListCallBuilder
Retrieves ACL entries on the specified object.

ObjectAccessControlMethodsBuilder
A builder providing access to all methods supported on objectAccessControl resources. It is not used directly, but through the Storage hub.

ObjectAccessControlPatchCallBuilder
Updates an ACL entry on the specified object. This method supports patch semantics.

ObjectAccessControlProjectTeam
The project team associated with the entity, if any.

ObjectAccessControlUpdateCallBuilder
Updates an ACL entry on the specified object.

ObjectAccessControls
An access-control list.

ObjectComposeCallBuilder
Concatenates a list of existing objects into a new object in the same bucket.

ObjectCopyCallBuilder
Copies an object to a specified location. Optionally overrides metadata.

ObjectDeleteCallBuilder
Deletes an object and its metadata. Deletions are permanent if versioning is not enabled for the bucket, or if the generation parameter is used.

ObjectGetCallBuilder
Retrieves an object or its metadata.

ObjectInsertCallBuilder
Stores a new object and metadata.

ObjectListCallBuilder
Retrieves a list of objects matching the criteria.

ObjectMethodsBuilder
A builder providing access to all methods supported on object resources. It is not used directly, but through the Storage hub.

ObjectOwner
The owner of the object. This will always be the uploader of the object.

ObjectPatchCallBuilder
Updates an object's metadata. This method supports patch semantics.

ObjectUpdateCallBuilder
Updates an object's metadata.

ObjectWatchAllCallBuilder
Watch for changes on all objects in a bucket.

Objects
A list of objects.

Storage
Central instance to access all Storage related resource activities

Enums

Result
A universal result type used as return for all action method results.

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

ResponseResult
Identifies types which are used in API responses.

UnusedType
Identifies types which are not actually used by the API This might be a bug within the google API schema.