Module google_storage1::api

Structs

Bucket
A bucket.

BucketAccessControl
An access-control entry.

BucketAccessControlDeleteCall
Permanently deletes the ACL entry for the specified entity on the specified bucket.

BucketAccessControlGetCall
Returns the ACL entry for the specified entity on the specified bucket.

BucketAccessControlInsertCall
Creates a new ACL entry on the specified bucket.

BucketAccessControlListCall
Retrieves ACL entries on the specified bucket.

BucketAccessControlMethods
A builder providing access to all methods supported on bucketAccessControl resources. It is not used directly, but through the Storage hub.

BucketAccessControlPatchCall
Patches an ACL entry on the specified bucket.

BucketAccessControlProjectTeam
The project team associated with the entity, if any.

BucketAccessControlUpdateCall
Updates an ACL entry on the specified bucket.

BucketAccessControls
An access-control list.

BucketBilling
The bucket's billing configuration.

BucketCors
The bucket's Cross-Origin Resource Sharing (CORS) configuration.

BucketDeleteCall
Permanently deletes an empty bucket.

BucketEncryption
Encryption configuration for a bucket.

BucketGetCall
Returns metadata for the specified bucket.

BucketGetIamPolicyCall
Returns an IAM policy for the specified bucket.

BucketIamConfiguration
The bucket's IAM configuration.

BucketIamConfigurationBucketPolicyOnly
The bucket's uniform bucket-level access configuration. The feature was formerly known as Bucket Policy Only. For backward compatibility, this field will be populated with identical information as the uniformBucketLevelAccess field. We recommend using the uniformBucketLevelAccess field to enable and disable the feature.

BucketIamConfigurationUniformBucketLevelAccess
The bucket's uniform bucket-level access configuration.

BucketInsertCall
Creates a new bucket.

BucketLifecycle
The bucket's lifecycle configuration. See lifecycle management for more information.

BucketLifecycleRule
A lifecycle management rule, which is made of an action to take and the condition(s) under which the action will be taken.

BucketLifecycleRuleAction
The action to take.

BucketLifecycleRuleCondition
The condition(s) under which the action will be taken.

BucketListCall
Retrieves a list of buckets for a given project.

BucketLockRetentionPolicyCall
Locks retention policy on a bucket.

BucketLogging
The bucket's logging configuration, which defines the destination bucket and optional name prefix for the current bucket's logs.

BucketMethods
A builder providing access to all methods supported on bucket resources. It is not used directly, but through the Storage hub.

BucketOwner
The owner of the bucket. This is always the project team's owner group.

BucketPatchCall
Patches a bucket. Changes to the bucket will be readable immediately after writing, but configuration changes may take time to propagate.

BucketRetentionPolicy
The bucket's retention policy. The retention policy enforces a minimum retention time for all objects contained in the bucket, based on their creation time. Any attempt to overwrite or delete objects younger than the retention period will result in a PERMISSION_DENIED error. An unlocked retention policy can be modified or removed from the bucket via a storage.buckets.update operation. A locked retention policy cannot be removed or shortened in duration for the lifetime of the bucket. Attempting to remove or decrease period of a locked retention policy will result in a PERMISSION_DENIED error.

BucketSetIamPolicyCall
Updates an IAM policy for the specified bucket.

BucketTestIamPermissionCall
Tests a set of permissions on the given bucket to see which, if any, are held by the caller.

BucketUpdateCall
Updates a bucket. Changes to the bucket will be readable immediately after writing, but configuration changes may take time to propagate.

BucketVersioning
The bucket's versioning configuration.

BucketWebsite
The bucket's website configuration, controlling how the service behaves when accessing bucket contents as a web site. See the Static Website Examples for more information.

Buckets
A list of buckets.

Channel
An notification channel used to watch for resource changes.

ChannelMethods
A builder providing access to all methods supported on channel resources. It is not used directly, but through the Storage hub.

ChannelStopCall
Stop watching resources through this channel

ComposeRequest
A Compose request.

ComposeRequestSourceObjects
The list of source objects that will be concatenated into a single object.

ComposeRequestSourceObjectsObjectPreconditions
Conditions that must be met for this operation to execute.

DefaultObjectAccessControlDeleteCall
Permanently deletes the default object ACL entry for the specified entity on the specified bucket.

DefaultObjectAccessControlGetCall
Returns the default object ACL entry for the specified entity on the specified bucket.

DefaultObjectAccessControlInsertCall
Creates a new default object ACL entry on the specified bucket.

DefaultObjectAccessControlListCall
Retrieves default object ACL entries on the specified bucket.

DefaultObjectAccessControlMethods
A builder providing access to all methods supported on defaultObjectAccessControl resources. It is not used directly, but through the Storage hub.

DefaultObjectAccessControlPatchCall
Patches a default object ACL entry on the specified bucket.

DefaultObjectAccessControlUpdateCall
Updates a default object ACL entry on the specified bucket.

Expr
Represents an expression text. Example: title: "User account presence" description: "Determines whether the request has a user account" expression: "size(request.user) > 0"

HmacKey
JSON template to produce a JSON-style HMAC Key resource for Create responses.

HmacKeyMetadata
JSON template to produce a JSON-style HMAC Key metadata resource.

HmacKeysMetadata
A list of hmacKeys.

Notification
A subscription to receive Google PubSub notifications.

NotificationDeleteCall
Permanently deletes a notification subscription.

NotificationGetCall
View a notification configuration.

NotificationInsertCall
Creates a notification subscription for a given bucket.

NotificationListCall
Retrieves a list of notification subscriptions for a given bucket.

NotificationMethods
A builder providing access to all methods supported on notification resources. It is not used directly, but through the Storage hub.

Notifications
A list of notification subscriptions.

Object
An object.

ObjectAccessControl
An access-control entry.

ObjectAccessControlDeleteCall
Permanently deletes the ACL entry for the specified entity on the specified object.

ObjectAccessControlGetCall
Returns the ACL entry for the specified entity on the specified object.

ObjectAccessControlInsertCall
Creates a new ACL entry on the specified object.

ObjectAccessControlListCall
Retrieves ACL entries on the specified object.

ObjectAccessControlMethods
A builder providing access to all methods supported on objectAccessControl resources. It is not used directly, but through the Storage hub.

ObjectAccessControlPatchCall
Patches an ACL entry on the specified object.

ObjectAccessControlProjectTeam
The project team associated with the entity, if any.

ObjectAccessControlUpdateCall
Updates an ACL entry on the specified object.

ObjectAccessControls
An access-control list.

ObjectComposeCall
Concatenates a list of existing objects into a new object in the same bucket.

ObjectCopyCall
Copies a source object to a destination object. Optionally overrides metadata.

ObjectCustomerEncryption
Metadata of customer-supplied encryption key, if the object is encrypted by such a key.

ObjectDeleteCall
Deletes an object and its metadata. Deletions are permanent if versioning is not enabled for the bucket, or if the generation parameter is used.

ObjectGetCall
Retrieves an object or its metadata.

ObjectGetIamPolicyCall
Returns an IAM policy for the specified object.

ObjectInsertCall
Stores a new object and metadata.

ObjectListCall
Retrieves a list of objects matching the criteria.

ObjectMethods
A builder providing access to all methods supported on object resources. It is not used directly, but through the Storage hub.

ObjectOwner
The owner of the object. This will always be the uploader of the object.

ObjectPatchCall
Patches an object's metadata.

ObjectRewriteCall
Rewrites a source object to a destination object. Optionally overrides metadata.

ObjectSetIamPolicyCall
Updates an IAM policy for the specified object.

ObjectTestIamPermissionCall
Tests a set of permissions on the given object to see which, if any, are held by the caller.

ObjectUpdateCall
Updates an object's metadata.

ObjectWatchAllCall
Watch for changes on all objects in a bucket.

Objects
A list of objects.

Policy
A bucket/object IAM policy.

PolicyBindings
An association between a role, which comes with a set of permissions, and members who may assume that role.

ProjectHmacKeyCreateCall
Creates a new HMAC key for the specified service account.

ProjectHmacKeyDeleteCall
Deletes an HMAC key.

ProjectHmacKeyGetCall
Retrieves an HMAC key's metadata

ProjectHmacKeyListCall
Retrieves a list of HMAC keys matching the criteria.

ProjectHmacKeyUpdateCall
Updates the state of an HMAC key. See the HMAC Key resource descriptor for valid states.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the Storage hub.

ProjectServiceAccountGetCall
Get the email address of this project's Google Cloud Storage service account.

RewriteResponse
A rewrite response.

ServiceAccount
A subscription to receive Google PubSub notifications.

Storage
Central instance to access all Storage related resource activities

TestIamPermissionsResponse
A storage.(buckets|objects).testIamPermissions response.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.