Struct google_storage1::api::Bucket

pub struct Bucket {
    pub acl: Option<Vec<BucketAccessControl>>,
    pub billing: Option<BucketBilling>,
    pub cors: Option<Vec<BucketCors>>,
    pub default_event_based_hold: Option<bool>,
    pub default_object_acl: Option<Vec<ObjectAccessControl>>,
    pub encryption: Option<BucketEncryption>,
    pub etag: Option<String>,
    pub iam_configuration: Option<BucketIamConfiguration>,
    pub id: Option<String>,
    pub kind: Option<String>,
    pub labels: Option<HashMap<String, String>>,
    pub lifecycle: Option<BucketLifecycle>,
    pub location: Option<String>,
    pub location_type: Option<String>,
    pub logging: Option<BucketLogging>,
    pub metageneration: Option<String>,
    pub name: Option<String>,
    pub owner: Option<BucketOwner>,
    pub project_number: Option<String>,
    pub retention_policy: Option<BucketRetentionPolicy>,
    pub satisfies_pzs: Option<bool>,
    pub self_link: Option<String>,
    pub storage_class: Option<String>,
    pub time_created: Option<String>,
    pub updated: Option<String>,
    pub versioning: Option<BucketVersioning>,
    pub website: Option<BucketWebsite>,
    pub zone_affinity: Option<Vec<String>>,
}

A bucket.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

delete buckets (none)

get buckets (response)

get iam policy buckets (none)

insert buckets (request|response)

list buckets (none)

lock retention policy buckets (response)

patch buckets (request|response)

set iam policy buckets (none)

test iam permissions buckets (none)

update buckets (request|response)

Fields
acl: Option<Vec<BucketAccessControl>>

Access controls on the bucket.

billing: Option<BucketBilling>

The bucket's billing configuration.

cors: Option<Vec<BucketCors>>

The bucket's Cross-Origin Resource Sharing (CORS) configuration.

default_event_based_hold: Option<bool>

The default value for event-based hold on newly created objects in this bucket. Event-based hold is a way to retain objects indefinitely until an event occurs, signified by the hold's release. After being released, such objects will be subject to bucket-level retention (if any). One sample use case of this flag is for banks to hold loan documents for at least 3 years after loan is paid in full. Here, bucket-level retention is 3 years and the event is loan being paid in full. In this example, these objects will be held intact for any number of years until the event has occurred (event-based hold on the object is released) and then 3 more years after that. That means retention duration of the objects begins from the moment event-based hold transitioned from true to false. Objects under event-based hold cannot be deleted, overwritten or archived until the hold is removed.

default_object_acl: Option<Vec<ObjectAccessControl>>

Default access controls to apply to new objects when no ACL is provided.

encryption: Option<BucketEncryption>

Encryption configuration for a bucket.

etag: Option<String>

HTTP 1.1 Entity tag for the bucket.

iam_configuration: Option<BucketIamConfiguration>

The bucket's IAM configuration.

id: Option<String>

The ID of the bucket. For buckets, the id and name properties are the same.

kind: Option<String>

The kind of item this is. For buckets, this is always storage#bucket.

labels: Option<HashMap<String, String>>

User-provided labels, in key/value pairs.

lifecycle: Option<BucketLifecycle>

The bucket's lifecycle configuration. See lifecycle management for more information.

location: Option<String>

The location of the bucket. Object data for objects in the bucket resides in physical storage within this region. Defaults to US. See the developer's guide for the authoritative list.

location_type: Option<String>

The type of the bucket location.

logging: Option<BucketLogging>

The bucket's logging configuration, which defines the destination bucket and optional name prefix for the current bucket's logs.

metageneration: Option<String>

The metadata generation of this bucket.

name: Option<String>

The name of the bucket.

owner: Option<BucketOwner>

The owner of the bucket. This is always the project team's owner group.

project_number: Option<String>

The project number of the project the bucket belongs to.

retention_policy: Option<BucketRetentionPolicy>

The bucket's retention policy. The retention policy enforces a minimum retention time for all objects contained in the bucket, based on their creation time. Any attempt to overwrite or delete objects younger than the retention period will result in a PERMISSION_DENIED error. An unlocked retention policy can be modified or removed from the bucket via a storage.buckets.update operation. A locked retention policy cannot be removed or shortened in duration for the lifetime of the bucket. Attempting to remove or decrease period of a locked retention policy will result in a PERMISSION_DENIED error.

satisfies_pzs: Option<bool>

Reserved for future use.

self_link: Option<String>

The URI of this bucket.

storage_class: Option<String>

The bucket's default storage class, used whenever no storageClass is specified for a newly-created object. This defines how objects in the bucket are stored and determines the SLA and the cost of storage. Values include MULTI_REGIONAL, REGIONAL, STANDARD, NEARLINE, COLDLINE, ARCHIVE, and DURABLE_REDUCED_AVAILABILITY. If this value is not specified when the bucket is created, it will default to STANDARD. For more information, see storage classes.

time_created: Option<String>

The creation time of the bucket in RFC 3339 format.

updated: Option<String>

The modification time of the bucket in RFC 3339 format.

versioning: Option<BucketVersioning>

The bucket's versioning configuration.

website: Option<BucketWebsite>

The bucket's website configuration, controlling how the service behaves when accessing bucket contents as a web site. See the Static Website Examples for more information.

zone_affinity: Option<Vec<String>>

The zone or zones from which the bucket is intended to use zonal quota. Requests for data from outside the specified affinities are still allowed but won't be able to use zonal quota. The zone or zones need to be within the bucket location otherwise the requests will fail with a 400 Bad Request response.