Struct google_storage1::api::ProjectMethods

pub struct ProjectMethods<'a, C> where
    C: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on project resources. It is not used directly, but through the Storage hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_storage1 as storage1;
 
use std::default::Default;
use oauth2;
use storage1::Storage;
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = Storage::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `hmac_keys_create(...)`, `hmac_keys_delete(...)`, `hmac_keys_get(...)`, `hmac_keys_list(...)`, `hmac_keys_update(...)` and `service_account_get(...)`
// to build up your call.
let rb = hub.projects();

Implementations

impl<'a, C> ProjectMethods<'a, C>

pub fn hmac_keys_create(
&self,
project_id: &str,
service_account_email: &str
) -> ProjectHmacKeyCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new HMAC key for the specified service account.

Arguments

projectId - Project ID owning the service account.

serviceAccountEmail - Email address of the service account.

pub fn hmac_keys_delete(
&self,
project_id: &str,
access_id: &str
) -> ProjectHmacKeyDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes an HMAC key.

Arguments

projectId - Project ID owning the requested key

accessId - Name of the HMAC key to be deleted.

pub fn hmac_keys_get(
&self,
project_id: &str,
access_id: &str
) -> ProjectHmacKeyGetCall<'a, C>

Create a builder to help you perform the following task:

Retrieves an HMAC key's metadata

Arguments

projectId - Project ID owning the service account of the requested key.

accessId - Name of the HMAC key.

pub fn hmac_keys_list(&self, project_id: &str) -> ProjectHmacKeyListCall<'a, C>

Create a builder to help you perform the following task:

Retrieves a list of HMAC keys matching the criteria.

Arguments

projectId - Name of the project in which to look for HMAC keys.

pub fn hmac_keys_update(
&self,
request: HmacKeyMetadata,
project_id: &str,
access_id: &str
) -> ProjectHmacKeyUpdateCall<'a, C>

Create a builder to help you perform the following task:

Updates the state of an HMAC key. See the HMAC Key resource descriptor for valid states.

Arguments

request - No description provided.

projectId - Project ID owning the service account of the updated key.

accessId - Name of the HMAC key being updated.

pub fn service_account_get(
&self,
project_id: &str
) -> ProjectServiceAccountGetCall<'a, C>

Create a builder to help you perform the following task:

Get the email address of this project's Google Cloud Storage service account.

Arguments

projectId - Project ID