Struct google_storage1::Bucket

pub struct Bucket {
    pub website: Option<BucketWebsite>,
    pub self_link: Option<String>,
    pub updated: Option<String>,
    pub iam_configuration: Option<BucketIamConfiguration>,
    pub time_created: Option<String>,
    pub labels: Option<HashMap<String, String>>,
    pub default_object_acl: Option<Vec<ObjectAccessControl>>,
    pub retention_policy: Option<BucketRetentionPolicy>,
    pub cors: Option<Vec<BucketCors>>,
    pub owner: Option<BucketOwner>,
    pub logging: Option<BucketLogging>,
    pub acl: Option<Vec<BucketAccessControl>>,
    pub id: Option<String>,
    pub project_number: Option<String>,
    pub metageneration: Option<String>,
    pub kind: Option<String>,
    pub default_event_based_hold: Option<bool>,
    pub name: Option<String>,
    pub billing: Option<BucketBilling>,
    pub encryption: Option<BucketEncryption>,
    pub lifecycle: Option<BucketLifecycle>,
    pub etag: Option<String>,
    pub location: Option<String>,
    pub location_type: Option<String>,
    pub versioning: Option<BucketVersioning>,
    pub storage_class: Option<String>,
}

A bucket.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

lock retention policy buckets (response)

update buckets (request|response)

test iam permissions buckets (none)

insert buckets (request|response)

delete buckets (none)

set iam policy buckets (none)

get iam policy buckets (none)

patch buckets (request|response)

list buckets (none)

get buckets (response)

Fields
website: Option<BucketWebsite>

The bucket's website configuration, controlling how the service behaves when accessing bucket contents as a web site. See the Static Website Examples for more information.

self_link: Option<String>

The URI of this bucket.

updated: Option<String>

The modification time of the bucket in RFC 3339 format.

iam_configuration: Option<BucketIamConfiguration>

The bucket's IAM configuration.

time_created: Option<String>

The creation time of the bucket in RFC 3339 format.

labels: Option<HashMap<String, String>>

User-provided labels, in key/value pairs.

default_object_acl: Option<Vec<ObjectAccessControl>>

Default access controls to apply to new objects when no ACL is provided.

retention_policy: Option<BucketRetentionPolicy>

The bucket's retention policy. The retention policy enforces a minimum retention time for all objects contained in the bucket, based on their creation time. Any attempt to overwrite or delete objects younger than the retention period will result in a PERMISSION_DENIED error. An unlocked retention policy can be modified or removed from the bucket via a storage.buckets.update operation. A locked retention policy cannot be removed or shortened in duration for the lifetime of the bucket. Attempting to remove or decrease period of a locked retention policy will result in a PERMISSION_DENIED error.

cors: Option<Vec<BucketCors>>

The bucket's Cross-Origin Resource Sharing (CORS) configuration.

owner: Option<BucketOwner>

The owner of the bucket. This is always the project team's owner group.

logging: Option<BucketLogging>

The bucket's logging configuration, which defines the destination bucket and optional name prefix for the current bucket's logs.

acl: Option<Vec<BucketAccessControl>>

Access controls on the bucket.

id: Option<String>

The ID of the bucket. For buckets, the id and name properties are the same.

project_number: Option<String>

The project number of the project the bucket belongs to.

metageneration: Option<String>

The metadata generation of this bucket.

kind: Option<String>

The kind of item this is. For buckets, this is always storage#bucket.

default_event_based_hold: Option<bool>

The default value for event-based hold on newly created objects in this bucket. Event-based hold is a way to retain objects indefinitely until an event occurs, signified by the hold's release. After being released, such objects will be subject to bucket-level retention (if any). One sample use case of this flag is for banks to hold loan documents for at least 3 years after loan is paid in full. Here, bucket-level retention is 3 years and the event is loan being paid in full. In this example, these objects will be held intact for any number of years until the event has occurred (event-based hold on the object is released) and then 3 more years after that. That means retention duration of the objects begins from the moment event-based hold transitioned from true to false. Objects under event-based hold cannot be deleted, overwritten or archived until the hold is removed.

name: Option<String>

The name of the bucket.

billing: Option<BucketBilling>

The bucket's billing configuration.

encryption: Option<BucketEncryption>

Encryption configuration for a bucket.

lifecycle: Option<BucketLifecycle>

The bucket's lifecycle configuration. See lifecycle management for more information.

etag: Option<String>

HTTP 1.1 Entity tag for the bucket.

location: Option<String>

The location of the bucket. Object data for objects in the bucket resides in physical storage within this region. Defaults to US. See the developer's guide for the authoritative list.

location_type: Option<String>

The type of the bucket location.

versioning: Option<BucketVersioning>

The bucket's versioning configuration.

storage_class: Option<String>

The bucket's default storage class, used whenever no storageClass is specified for a newly-created object. This defines how objects in the bucket are stored and determines the SLA and the cost of storage. Values include MULTI_REGIONAL, REGIONAL, STANDARD, NEARLINE, COLDLINE, and DURABLE_REDUCED_AVAILABILITY. If this value is not specified when the bucket is created, it will default to STANDARD. For more information, see storage classes.