Struct google_storage1::BucketAccessControlMethods

pub struct BucketAccessControlMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on bucketAccessControl resources. It is not used directly, but through the Storage hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_storage1 as storage1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use storage1::Storage;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Storage::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `delete(...)`, `get(...)`, `insert(...)`, `list(...)`, `patch(...)` and `update(...)`
// to build up your call.
let rb = hub.bucket_access_controls();

Methods

impl<'a, C, A> BucketAccessControlMethods<'a, C, A>

fn patch(&self, request: BucketAccessControl, bucket: &str, entity: &str) -> BucketAccessControlPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an ACL entry on the specified bucket. This method supports patch semantics.

Arguments

request - No description provided.

bucket - Name of a bucket.

entity - The entity holding the permission. Can be user-userId, user-emailAddress, group-groupId, group-emailAddress, allUsers, or allAuthenticatedUsers.

fn delete(&self, bucket: &str, entity: &str) -> BucketAccessControlDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Permanently deletes the ACL entry for the specified entity on the specified bucket.

Arguments

bucket - Name of a bucket.

entity - The entity holding the permission. Can be user-userId, user-emailAddress, group-groupId, group-emailAddress, allUsers, or allAuthenticatedUsers.

fn insert(&self, request: BucketAccessControl, bucket: &str) -> BucketAccessControlInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new ACL entry on the specified bucket.

Arguments

request - No description provided.

bucket - Name of a bucket.

fn get(&self, bucket: &str, entity: &str) -> BucketAccessControlGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns the ACL entry for the specified entity on the specified bucket.

Arguments

bucket - Name of a bucket.

entity - The entity holding the permission. Can be user-userId, user-emailAddress, group-groupId, group-emailAddress, allUsers, or allAuthenticatedUsers.

fn update(&self, request: BucketAccessControl, bucket: &str, entity: &str) -> BucketAccessControlUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an ACL entry on the specified bucket.

Arguments

request - No description provided.

bucket - Name of a bucket.

entity - The entity holding the permission. Can be user-userId, user-emailAddress, group-groupId, group-emailAddress, allUsers, or allAuthenticatedUsers.

fn list(&self, bucket: &str) -> BucketAccessControlListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves ACL entries on the specified bucket.

Arguments

bucket - Name of a bucket.