Struct google_storage1::BucketUpdateCall

pub struct BucketUpdateCall<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

Updates a bucket.

A builder for the update method supported by a bucket resource. It is not used directly, but through a BucketMethods instance.

Example

Instantiate a resource method builder

use storage1::Bucket;
 
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = Bucket::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.buckets().update(req, "bucket")
             .projection("voluptua.")
             .predefined_default_object_acl("dolor")
             .predefined_acl("et")
             .if_metageneration_not_match("et")
             .if_metageneration_match("vero")
             .doit();

Methods

impl<'a, C, A> BucketUpdateCall<'a, C, A> where C: BorrowMut<Client>, A: GetToken

fn doit(self) -> Result<(Response, Bucket)>

Perform the operation you have build so far.

fn request(self, new_value: Bucket) -> BucketUpdateCall<'a, C, A>

Sets the request property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

fn bucket(self, new_value: &str) -> BucketUpdateCall<'a, C, A>

Name of a bucket.

Sets the bucket path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

fn projection(self, new_value: &str) -> BucketUpdateCall<'a, C, A>

Set of properties to return. Defaults to full.

Sets the projection query property to the given value.

fn predefined_default_object_acl(self, new_value: &str) -> BucketUpdateCall<'a, C, A>

Apply a predefined set of default object access controls to this bucket.

Sets the predefined default object acl query property to the given value.

fn predefined_acl(self, new_value: &str) -> BucketUpdateCall<'a, C, A>

Apply a predefined set of access controls to this bucket.

Sets the predefined acl query property to the given value.

fn if_metageneration_not_match(self, new_value: &str) -> BucketUpdateCall<'a, C, A>

Makes the return of the bucket metadata conditional on whether the bucket's current metageneration does not match the given value.

Sets the if metageneration not match query property to the given value.

fn if_metageneration_match(self, new_value: &str) -> BucketUpdateCall<'a, C, A>

Makes the return of the bucket metadata conditional on whether the bucket's current metageneration matches the given value.

Sets the if metageneration match query property to the given value.

fn delegate(self, new_value: &'a mut Delegate) -> BucketUpdateCall<'a, C, A>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

fn param<T>(self, name: T, value: T) -> BucketUpdateCall<'a, C, A> where T: AsRef<str>

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known paramters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters. Overrides userIp if both are provided.

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

userIp (query-string) - IP address of the site where the request originates. Use this if you want to enforce per-user limits.

fields (query-string) - Selector specifying which fields to include in a partial response.

alt (query-string) - Data format for the response.

fn add_scope<T>(self, scope: T) -> BucketUpdateCall<'a, C, A> where T: AsRef<str>

Identifies the authorization scope for the method you are building.

Use this method to actively specify which scope should be used, instead the default Scope variant Scope::CloudPlatform.

The scope will be added to a set of scopes. This is important as one can maintain access tokens for more than one scope.

Usually there is more than one suitable scope to authorize an operation, some of which may encompass more rights than others. For example, for listing resources, a read-only scope will be sufficient, a read-write scope will do as well.