Struct google_storage1::ObjectAccessControlMethods

pub struct ObjectAccessControlMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on objectAccessControl resources. It is not used directly, but through the Storage hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_storage1 as storage1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use storage1::Storage;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Storage::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `delete(...)`, `get(...)`, `insert(...)`, `list(...)`, `patch(...)` and `update(...)`
// to build up your call.
let rb = hub.object_access_controls();

Methods

impl<'a, C, A> ObjectAccessControlMethods<'a, C, A>

fn insert(&self, request: ObjectAccessControl, bucket: &str, object: &str) -> ObjectAccessControlInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new ACL entry on the specified object.

Arguments

request - No description provided.

bucket - Name of a bucket.

object - Name of the object. For information about how to URL encode object names to be path safe, see Encoding URI Path Parts.

fn patch(&self, request: ObjectAccessControl, bucket: &str, object: &str, entity: &str) -> ObjectAccessControlPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an ACL entry on the specified object. This method supports patch semantics.

Arguments

request - No description provided.

bucket - Name of a bucket.

object - Name of the object. For information about how to URL encode object names to be path safe, see Encoding URI Path Parts.

entity - The entity holding the permission. Can be user-userId, user-emailAddress, group-groupId, group-emailAddress, allUsers, or allAuthenticatedUsers.

fn list(&self, bucket: &str, object: &str) -> ObjectAccessControlListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves ACL entries on the specified object.

Arguments

bucket - Name of a bucket.

object - Name of the object. For information about how to URL encode object names to be path safe, see Encoding URI Path Parts.

fn delete(&self, bucket: &str, object: &str, entity: &str) -> ObjectAccessControlDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Permanently deletes the ACL entry for the specified entity on the specified object.

Arguments

bucket - Name of a bucket.

object - Name of the object. For information about how to URL encode object names to be path safe, see Encoding URI Path Parts.

entity - The entity holding the permission. Can be user-userId, user-emailAddress, group-groupId, group-emailAddress, allUsers, or allAuthenticatedUsers.

fn get(&self, bucket: &str, object: &str, entity: &str) -> ObjectAccessControlGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns the ACL entry for the specified entity on the specified object.

Arguments

bucket - Name of a bucket.

object - Name of the object. For information about how to URL encode object names to be path safe, see Encoding URI Path Parts.

entity - The entity holding the permission. Can be user-userId, user-emailAddress, group-groupId, group-emailAddress, allUsers, or allAuthenticatedUsers.

fn update(&self, request: ObjectAccessControl, bucket: &str, object: &str, entity: &str) -> ObjectAccessControlUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an ACL entry on the specified object.

Arguments

request - No description provided.

bucket - Name of a bucket.

object - Name of the object. For information about how to URL encode object names to be path safe, see Encoding URI Path Parts.

entity - The entity holding the permission. Can be user-userId, user-emailAddress, group-groupId, group-emailAddress, allUsers, or allAuthenticatedUsers.