Struct google_storage1::ObjectInsertCall

pub struct ObjectInsertCall<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

Stores a new object and metadata.

This method supports media download. To enable it, adjust the builder like this: .param("alt", "media"). Please note that due to missing multi-part support on the server side, you will only receive the media, but not the Object structure that you would usually get. The latter will be a default value.

A builder for the insert method supported by a object resource. It is not used directly, but through a ObjectMethods instance.

Example

Instantiate a resource method builder

use storage1::Object;
use std::fs;
 
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = Object::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `upload(...)`.
// Values shown here are possibly random and not representative !
let result = hub.objects().insert(req, "bucket")
             .projection("rebum.")
             .predefined_acl("dolore")
             .name("nonumy")
             .if_metageneration_not_match("sed")
             .if_metageneration_match("aliquyam")
             .if_generation_not_match("sit")
             .if_generation_match("eirmod")
             .content_encoding("consetetur")
             .upload(fs::File::open("file.ext").unwrap(), "application/octet-stream".parse().unwrap());

Methods

impl<'a, C, A> ObjectInsertCall<'a, C, A> where C: BorrowMut<Client>, A: GetToken

fn upload<RS>(self, stream: RS, mime_type: Mime) -> Result<(Response, Object)> where RS: ReadSeek

Upload media all at once. If the upload fails for whichever reason, all progress is lost.

max size: 0kb

multipart: yes

valid mime types: '*/*'

fn upload_resumable<RS>(self, resumeable_stream: RS, mime_type: Mime) -> Result<(Response, Object)> where RS: ReadSeek

Upload media in a resumable fashion. Even if the upload fails or is interrupted, it can be resumed for a certain amount of time as the server maintains state temporarily.

The delegate will be asked for an upload_url(), and if not provided, will be asked to store an upload URL that was provided by the server, using store_upload_url(...). The upload will be done in chunks, the delegate may specify the chunk_size() and may cancel the operation before each chunk is uploaded, using cancel_chunk_upload(...).

max size: 0kb

multipart: yes

valid mime types: '*/*'

fn request(self, new_value: Object) -> ObjectInsertCall<'a, C, A>

Sets the request property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

fn bucket(self, new_value: &str) -> ObjectInsertCall<'a, C, A>

Name of the bucket in which to store the new object. Overrides the provided object metadata's bucket value, if any.

Sets the bucket path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

fn projection(self, new_value: &str) -> ObjectInsertCall<'a, C, A>

Set of properties to return. Defaults to noAcl, unless the object resource specifies the acl property, when it defaults to full.

Sets the projection query property to the given value.

fn predefined_acl(self, new_value: &str) -> ObjectInsertCall<'a, C, A>

Apply a predefined set of access controls to this object.

Sets the predefined acl query property to the given value.

fn name(self, new_value: &str) -> ObjectInsertCall<'a, C, A>

Name of the object. Required when the object metadata is not otherwise provided. Overrides the object metadata's name value, if any. For information about how to URL encode object names to be path safe, see Encoding URI Path Parts.

Sets the name query property to the given value.

fn if_metageneration_not_match(self, new_value: &str) -> ObjectInsertCall<'a, C, A>

Makes the operation conditional on whether the object's current metageneration does not match the given value.

Sets the if metageneration not match query property to the given value.

fn if_metageneration_match(self, new_value: &str) -> ObjectInsertCall<'a, C, A>

Makes the operation conditional on whether the object's current metageneration matches the given value.

Sets the if metageneration match query property to the given value.

fn if_generation_not_match(self, new_value: &str) -> ObjectInsertCall<'a, C, A>

Makes the operation conditional on whether the object's current generation does not match the given value.

Sets the if generation not match query property to the given value.

fn if_generation_match(self, new_value: &str) -> ObjectInsertCall<'a, C, A>

Makes the operation conditional on whether the object's current generation matches the given value.

Sets the if generation match query property to the given value.

fn content_encoding(self, new_value: &str) -> ObjectInsertCall<'a, C, A>

If set, sets the contentEncoding property of the final object to this value. Setting this parameter is equivalent to setting the contentEncoding metadata property. This can be useful when uploading an object with uploadType=media to indicate the encoding of the content being uploaded.

Sets the content encoding query property to the given value.

fn delegate(self, new_value: &'a mut Delegate) -> ObjectInsertCall<'a, C, A>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

fn param<T>(self, name: T, value: T) -> ObjectInsertCall<'a, C, A> where T: AsRef<str>

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known paramters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters. Overrides userIp if both are provided.

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

userIp (query-string) - IP address of the site where the request originates. Use this if you want to enforce per-user limits.

fields (query-string) - Selector specifying which fields to include in a partial response.

alt (query-string) - Data format for the response.

fn add_scope<T>(self, scope: T) -> ObjectInsertCall<'a, C, A> where T: AsRef<str>

Identifies the authorization scope for the method you are building.

Use this method to actively specify which scope should be used, instead the default Scope variant Scope::CloudPlatform.

The scope will be added to a set of scopes. This is important as one can maintain access tokens for more than one scope.

Usually there is more than one suitable scope to authorize an operation, some of which may encompass more rights than others. For example, for listing resources, a read-only scope will be sufficient, a read-write scope will do as well.