Struct google_storage1::ObjectMethods

pub struct ObjectMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on object resources. It is not used directly, but through the Storage hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_storage1 as storage1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use storage1::Storage;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Storage::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `compose(...)`, `copy(...)`, `delete(...)`, `get(...)`, `insert(...)`, `list(...)`, `patch(...)`, `rewrite(...)`, `update(...)` and `watch_all(...)`
// to build up your call.
let rb = hub.objects();

Methods

impl<'a, C, A> ObjectMethods<'a, C, A>

fn rewrite(&self, request: Object, source_bucket: &str, source_object: &str, destination_bucket: &str, destination_object: &str) -> ObjectRewriteCall<'a, C, A>

Create a builder to help you perform the following task:

Rewrites a source object to a destination object. Optionally overrides metadata.

Arguments

request - No description provided.

sourceBucket - Name of the bucket in which to find the source object.

sourceObject - Name of the source object.

destinationBucket - Name of the bucket in which to store the new object. Overrides the provided object metadata's bucket value, if any.

destinationObject - Name of the new object. Required when the object metadata is not otherwise provided. Overrides the object metadata's name value, if any.

fn get(&self, bucket: &str, object: &str) -> ObjectGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves an object or its metadata.

Arguments

bucket - Name of the bucket in which the object resides.

object - Name of the object.

fn update(&self, request: Object, bucket: &str, object: &str) -> ObjectUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an object's metadata.

Arguments

request - No description provided.

bucket - Name of the bucket in which the object resides.

object - Name of the object.

fn watch_all(&self, request: Channel, bucket: &str) -> ObjectWatchAllCall<'a, C, A>

Create a builder to help you perform the following task:

Watch for changes on all objects in a bucket.

Arguments

request - No description provided.

bucket - Name of the bucket in which to look for objects.

fn insert(&self, request: Object, bucket: &str) -> ObjectInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Stores a new object and metadata.

Arguments

request - No description provided.

bucket - Name of the bucket in which to store the new object. Overrides the provided object metadata's bucket value, if any.

fn compose(&self, request: ComposeRequest, destination_bucket: &str, destination_object: &str) -> ObjectComposeCall<'a, C, A>

Create a builder to help you perform the following task:

Concatenates a list of existing objects into a new object in the same bucket.

Arguments

request - No description provided.

destinationBucket - Name of the bucket in which to store the new object.

destinationObject - Name of the new object.

fn delete(&self, bucket: &str, object: &str) -> ObjectDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes an object and its metadata. Deletions are permanent if versioning is not enabled for the bucket, or if the generation parameter is used.

Arguments

bucket - Name of the bucket in which the object resides.

object - Name of the object.

fn list(&self, bucket: &str) -> ObjectListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves a list of objects matching the criteria.

Arguments

bucket - Name of the bucket in which to look for objects.

fn copy(&self, request: Object, source_bucket: &str, source_object: &str, destination_bucket: &str, destination_object: &str) -> ObjectCopyCall<'a, C, A>

Create a builder to help you perform the following task:

Copies an object to a specified location. Optionally overrides metadata.

Arguments

request - No description provided.

sourceBucket - Name of the bucket in which to find the source object.

sourceObject - Name of the source object.

destinationBucket - Name of the bucket in which to store the new object. Overrides the provided object metadata's bucket value, if any.

destinationObject - Name of the new object. Required when the object metadata is not otherwise provided. Overrides the object metadata's name value, if any.

fn patch(&self, request: Object, bucket: &str, object: &str) -> ObjectPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an object's metadata. This method supports patch semantics.

Arguments

request - No description provided.

bucket - Name of the bucket in which the object resides.

object - Name of the object.