Struct google-storage1::Storage

pub struct Storage<C, NC, A> {
    // some fields omitted
}

Central instance to access all Storage related resource activities

Examples

Instantiate a new hub

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "google-storage1" as storage1;
use storage1::ObjectAccessControl;
use storage1::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use storage1::Storage;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Storage::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req: ObjectAccessControl = Default::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.object_access_controls().update(&req, "bucket", "object", "entity")
             .generation("erat")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::MissingAPIKey => println!("Auth: Missing API Key - used if there are no scopes"),
    Result::MissingToken => println!("OAuth2: Missing Token"),
    Result::Cancelled => println!("Operation cancelled by user"),
    Result::UploadSizeLimitExceeded(size, max_size) => println!("Upload size too big: {} of {}", size, max_size),
    Result::Failure(_) => println!("General Failure (hyper::client::Response doesn't print)"),
    Result::FieldClash(clashed_field) => println!("You added custom parameter which is part of builder: {:?}", clashed_field),
    Result::JsonDecodeError(err) => println!("Couldn't understand server reply - maybe API needs update: {:?}", err),
    Result::Success(_) => println!("Success (value doesn't print)"),
}

Methods

impl<'a, C, NC, A> Storage<C, NC, A> where NC: NetworkConnector, C: BorrowMut<Client<NC>>, A: GetToken

fn new(client: C, authenticator: A) -> Storage<C, NC, A>

fn bucket_access_controls(&'a self) -> BucketAccessControlMethods<'a, C, NC, A>

fn buckets(&'a self) -> BucketMethods<'a, C, NC, A>

fn channels(&'a self) -> ChannelMethods<'a, C, NC, A>

fn default_object_access_controls(&'a self) -> DefaultObjectAccessControlMethods<'a, C, NC, A>

fn object_access_controls(&'a self) -> ObjectAccessControlMethods<'a, C, NC, A>

fn objects(&'a self) -> ObjectMethods<'a, C, NC, A>

fn user_agent(&mut self, agent_name: String) -> String

Set the user-agent header field to use in all requests to the server. It defaults to google-api-rust-client/0.1.0.

Returns the previously set user-agent.