Crate google_storage1

This documentation was generated from storage crate version 1.0.9+20190624, where 20190624 is the exact revision of the storage:v1 schema built by the mako code generator v1.0.9.

Everything else about the storage v1 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

bucket access controls

delete, get, insert, list, patch and update

buckets

delete, get, get iam policy, insert, list, lock retention policy, patch, set iam policy, test iam permissions and update

channels

stop

default object access controls

delete, get, insert, list, patch and update

notifications

delete, get, insert and list

object access controls

delete, get, insert, list, patch and update

objects

compose, copy, delete, get, get iam policy, insert, list, patch, rewrite, set iam policy, test iam permissions, update and watch all

projects

hmac keys create, hmac keys delete, hmac keys get, hmac keys list, hmac keys update and service account get

Upload supported by ...

insert objects

Download supported by ...

get objects

Subscription supported by ...

watch all objects

list objects

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.objects().get_iam_policy(...).doit()
let r = hub.objects().list(...).doit()
let r = hub.objects().rewrite(...).doit()
let r = hub.objects().test_iam_permissions(...).doit()
let r = hub.objects().copy(...).doit()
let r = hub.objects().watch_all(...).doit()
let r = hub.objects().get(...).doit()
let r = hub.objects().set_iam_policy(...).doit()
let r = hub.objects().insert(...).doit()
let r = hub.objects().compose(...).doit()
let r = hub.objects().update(...).doit()
let r = hub.objects().delete(...).doit()
let r = hub.objects().patch(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-storage1 = "*"
# This project intentionally uses an old version of Hyper. See
# https://github.com/Byron/google-apis-rs/issues/173 for more
# information.
hyper = "^0.10"
hyper-rustls = "^0.6"
serde = "^1.0"
serde_json = "^1.0"
yup-oauth2 = "^1.0"

A complete example

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_storage1 as storage1;
use storage1::Object;
use storage1::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use storage1::Storage;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Storage::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = Object::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.objects().rewrite(req, "sourceBucket", "sourceObject", "destinationBucket", "destinationObject")
             .user_project("et")
             .source_generation("diam")
             .rewrite_token("ipsum")
             .provisional_user_project("Lorem")
             .projection("et")
             .max_bytes_rewritten_per_call("duo")
             .if_source_metageneration_not_match("aliquyam")
             .if_source_metageneration_match("sea")
             .if_source_generation_not_match("Lorem")
             .if_source_generation_match("eos")
             .if_metageneration_not_match("erat")
             .if_metageneration_match("sadipscing")
             .if_generation_not_match("dolor")
             .if_generation_match("eirmod")
             .destination_predefined_acl("elitr")
             .destination_kms_key_name("amet")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

Bucket
A bucket.

BucketAccessControl
An access-control entry.

BucketAccessControlDeleteCall
Permanently deletes the ACL entry for the specified entity on the specified bucket.

BucketAccessControlGetCall
Returns the ACL entry for the specified entity on the specified bucket.

BucketAccessControlInsertCall
Creates a new ACL entry on the specified bucket.

BucketAccessControlListCall
Retrieves ACL entries on the specified bucket.

BucketAccessControlMethods
A builder providing access to all methods supported on bucketAccessControl resources. It is not used directly, but through the Storage hub.

BucketAccessControlPatchCall
Patches an ACL entry on the specified bucket.

BucketAccessControlProjectTeam
The project team associated with the entity, if any.

BucketAccessControlUpdateCall
Updates an ACL entry on the specified bucket.

BucketAccessControls
An access-control list.

BucketBilling
The bucket's billing configuration.

BucketCors
The bucket's Cross-Origin Resource Sharing (CORS) configuration.

BucketDeleteCall
Permanently deletes an empty bucket.

BucketEncryption
Encryption configuration for a bucket.

BucketGetCall
Returns metadata for the specified bucket.

BucketGetIamPolicyCall
Returns an IAM policy for the specified bucket.

BucketIamConfiguration
The bucket's IAM configuration.

BucketIamConfigurationBucketPolicyOnly
The bucket's Bucket Policy Only configuration.

BucketIamConfigurationUniformBucketLevelAccess
The bucket's uniform bucket-level access configuration.

BucketInsertCall
Creates a new bucket.

BucketLifecycle
The bucket's lifecycle configuration. See lifecycle management for more information.

BucketLifecycleRule
A lifecycle management rule, which is made of an action to take and the condition(s) under which the action will be taken.

BucketLifecycleRuleAction
The action to take.

BucketLifecycleRuleCondition
The condition(s) under which the action will be taken.

BucketListCall
Retrieves a list of buckets for a given project.

BucketLockRetentionPolicyCall
Locks retention policy on a bucket.

BucketLogging
The bucket's logging configuration, which defines the destination bucket and optional name prefix for the current bucket's logs.

BucketMethods
A builder providing access to all methods supported on bucket resources. It is not used directly, but through the Storage hub.

BucketOwner
The owner of the bucket. This is always the project team's owner group.

BucketPatchCall
Patches a bucket. Changes to the bucket will be readable immediately after writing, but configuration changes may take time to propagate.

BucketRetentionPolicy
The bucket's retention policy. The retention policy enforces a minimum retention time for all objects contained in the bucket, based on their creation time. Any attempt to overwrite or delete objects younger than the retention period will result in a PERMISSION_DENIED error. An unlocked retention policy can be modified or removed from the bucket via a storage.buckets.update operation. A locked retention policy cannot be removed or shortened in duration for the lifetime of the bucket. Attempting to remove or decrease period of a locked retention policy will result in a PERMISSION_DENIED error.

BucketSetIamPolicyCall
Updates an IAM policy for the specified bucket.

BucketTestIamPermissionCall
Tests a set of permissions on the given bucket to see which, if any, are held by the caller.

BucketUpdateCall
Updates a bucket. Changes to the bucket will be readable immediately after writing, but configuration changes may take time to propagate.

BucketVersioning
The bucket's versioning configuration.

BucketWebsite
The bucket's website configuration, controlling how the service behaves when accessing bucket contents as a web site. See the Static Website Examples for more information.

Buckets
A list of buckets.

Channel
An notification channel used to watch for resource changes.

ChannelMethods
A builder providing access to all methods supported on channel resources. It is not used directly, but through the Storage hub.

ChannelStopCall
Stop watching resources through this channel

Chunk

ComposeRequest
A Compose request.

ComposeRequestSourceObjects
The list of source objects that will be concatenated into a single object.

ComposeRequestSourceObjectsObjectPreconditions
Conditions that must be met for this operation to execute.

ContentRange
Implements the Content-Range header, for serialization only

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

DefaultObjectAccessControlDeleteCall
Permanently deletes the default object ACL entry for the specified entity on the specified bucket.

DefaultObjectAccessControlGetCall
Returns the default object ACL entry for the specified entity on the specified bucket.

DefaultObjectAccessControlInsertCall
Creates a new default object ACL entry on the specified bucket.

DefaultObjectAccessControlListCall
Retrieves default object ACL entries on the specified bucket.

DefaultObjectAccessControlMethods
A builder providing access to all methods supported on defaultObjectAccessControl resources. It is not used directly, but through the Storage hub.

DefaultObjectAccessControlPatchCall
Patches a default object ACL entry on the specified bucket.

DefaultObjectAccessControlUpdateCall
Updates a default object ACL entry on the specified bucket.

DummyNetworkStream

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

Expr
Represents an expression text. Example: title: "User account presence" description: "Determines whether the request has a user account" expression: "size(request.user) > 0"

HmacKey
JSON template to produce a JSON-style HMAC Key resource for Create responses.

HmacKeyMetadata
JSON template to produce a JSON-style HMAC Key metadata resource.

HmacKeysMetadata
A list of hmacKeys.

JsonServerError
A utility type which can decode a server response that indicates error

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

Notification
A subscription to receive Google PubSub notifications.

NotificationDeleteCall
Permanently deletes a notification subscription.

NotificationGetCall
View a notification configuration.

NotificationInsertCall
Creates a notification subscription for a given bucket.

NotificationListCall
Retrieves a list of notification subscriptions for a given bucket.

NotificationMethods
A builder providing access to all methods supported on notification resources. It is not used directly, but through the Storage hub.

Notifications
A list of notification subscriptions.

Object
An object.

ObjectAccessControl
An access-control entry.

ObjectAccessControlDeleteCall
Permanently deletes the ACL entry for the specified entity on the specified object.

ObjectAccessControlGetCall
Returns the ACL entry for the specified entity on the specified object.

ObjectAccessControlInsertCall
Creates a new ACL entry on the specified object.

ObjectAccessControlListCall
Retrieves ACL entries on the specified object.

ObjectAccessControlMethods
A builder providing access to all methods supported on objectAccessControl resources. It is not used directly, but through the Storage hub.

ObjectAccessControlPatchCall
Patches an ACL entry on the specified object.

ObjectAccessControlProjectTeam
The project team associated with the entity, if any.

ObjectAccessControlUpdateCall
Updates an ACL entry on the specified object.

ObjectAccessControls
An access-control list.

ObjectComposeCall
Concatenates a list of existing objects into a new object in the same bucket.

ObjectCopyCall
Copies a source object to a destination object. Optionally overrides metadata.

ObjectCustomerEncryption
Metadata of customer-supplied encryption key, if the object is encrypted by such a key.

ObjectDeleteCall
Deletes an object and its metadata. Deletions are permanent if versioning is not enabled for the bucket, or if the generation parameter is used.

ObjectGetCall
Retrieves an object or its metadata.

ObjectGetIamPolicyCall
Returns an IAM policy for the specified object.

ObjectInsertCall
Stores a new object and metadata.

ObjectListCall
Retrieves a list of objects matching the criteria.

ObjectMethods
A builder providing access to all methods supported on object resources. It is not used directly, but through the Storage hub.

ObjectOwner
The owner of the object. This will always be the uploader of the object.

ObjectPatchCall
Patches an object's metadata.

ObjectRewriteCall
Rewrites a source object to a destination object. Optionally overrides metadata.

ObjectSetIamPolicyCall
Updates an IAM policy for the specified object.

ObjectTestIamPermissionCall
Tests a set of permissions on the given object to see which, if any, are held by the caller.

ObjectUpdateCall
Updates an object's metadata.

ObjectWatchAllCall
Watch for changes on all objects in a bucket.

Objects
A list of objects.

Policy
A bucket/object IAM policy.

PolicyBindings
An association between a role, which comes with a set of permissions, and members who may assume that role.

ProjectHmacKeyCreateCall
Creates a new HMAC key for the specified service account.

ProjectHmacKeyDeleteCall
Deletes an HMAC key.

ProjectHmacKeyGetCall
Retrieves an HMAC key's metadata

ProjectHmacKeyListCall
Retrieves a list of HMAC keys matching the criteria.

ProjectHmacKeyUpdateCall
Updates the state of an HMAC key. See the HMAC Key resource descriptor for valid states.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the Storage hub.

ProjectServiceAccountGetCall
Get the email address of this project's Google Cloud Storage service account.

RangeResponseHeader

ResumableUploadHelper
A utility type to perform a resumable upload from start to end.

RewriteResponse
A rewrite response.

ServerError

ServerMessage

ServiceAccount
A subscription to receive Google PubSub notifications.

Storage
Central instance to access all Storage related resource activities

TestIamPermissionsResponse
A storage.(buckets|objects).testIamPermissions response.

XUploadContentType
The X-Upload-Content-Type header.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

UnusedType
Identifies types which are not actually used by the API This might be a bug within the google API schema.

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.