Struct google_storage1::Bucket

pub struct Bucket {
    pub website: Option<BucketWebsite>,
    pub time_created: Option<String>,
    pub versioning: Option<BucketVersioning>,
    pub default_object_acl: Option<Vec<ObjectAccessControl>>,
    pub metageneration: Option<String>,
    pub cors: Option<Vec<BucketCors>>,
    pub owner: Option<BucketOwner>,
    pub acl: Option<Vec<BucketAccessControl>>,
    pub id: Option<String>,
    pub kind: Option<String>,
    pub logging: Option<BucketLogging>,
    pub name: Option<String>,
    pub project_number: Option<String>,
    pub etag: Option<String>,
    pub storage_class: Option<String>,
    pub lifecycle: Option<BucketLifecycle>,
    pub self_link: Option<String>,
    pub location: Option<String>,
}

A bucket.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

update buckets (request|response)

insert buckets (request|response)

delete buckets (none)

patch buckets (request|response)

list buckets (none)

get buckets (response)

Fields

website

The bucket's website configuration.

time_created

Creation time of the bucket in RFC 3339 format.

versioning

The bucket's versioning configuration.

default_object_acl

Default access controls to apply to new objects when no ACL is provided.

metageneration

The metadata generation of this bucket.

cors

The bucket's Cross-Origin Resource Sharing (CORS) configuration.

owner

The owner of the bucket. This is always the project team's owner group.

acl

Access controls on the bucket.

id

The ID of the bucket.

kind

The kind of item this is. For buckets, this is always storage#bucket.

logging

The bucket's logging configuration, which defines the destination bucket and optional name prefix for the current bucket's logs.

name

The name of the bucket.

project_number

The project number of the project the bucket belongs to.

etag

HTTP 1.1 Entity tag for the bucket.

storage_class

The bucket's storage class. This defines how objects in the bucket are stored and determines the SLA and the cost of storage. Values include STANDARD, NEARLINE and DURABLE_REDUCED_AVAILABILITY. Defaults to STANDARD. For more information, see storage classes.

lifecycle

The bucket's lifecycle configuration. See lifecycle management for more information.

self_link

The URI of this bucket.

location

The location of the bucket. Object data for objects in the bucket resides in physical storage within this region. Defaults to US. See the developer's guide for the authoritative list.