Struct google_storage1::Bucket

pub struct Bucket {
    pub website: Option<BucketWebsite>,
    pub updated: Option<String>,
    pub time_created: Option<String>,
    pub versioning: Option<BucketVersioning>,
    pub default_object_acl: Option<Vec<ObjectAccessControl>>,
    pub metageneration: Option<String>,
    pub cors: Option<Vec<BucketCors>>,
    pub owner: Option<BucketOwner>,
    pub acl: Option<Vec<BucketAccessControl>>,
    pub id: Option<String>,
    pub kind: Option<String>,
    pub logging: Option<BucketLogging>,
    pub name: Option<String>,
    pub encryption: Option<BucketEncryption>,
    pub project_number: Option<String>,
    pub etag: Option<String>,
    pub storage_class: Option<String>,
    pub lifecycle: Option<BucketLifecycle>,
    pub self_link: Option<String>,
    pub location: Option<String>,
}

A bucket.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

update buckets (request|response)

insert buckets (request|response)

delete buckets (none)

patch buckets (request|response)

list buckets (none)

get buckets (response)

Fields
website: Option<BucketWebsite>

The bucket's website configuration, controlling how the service behaves when accessing bucket contents as a web site. See the Static Website Examples for more information.

updated: Option<String>

The modification time of the bucket in RFC 3339 format.

time_created: Option<String>

The creation time of the bucket in RFC 3339 format.

versioning: Option<BucketVersioning>

The bucket's versioning configuration.

default_object_acl: Option<Vec<ObjectAccessControl>>

Default access controls to apply to new objects when no ACL is provided.

metageneration: Option<String>

The metadata generation of this bucket.

cors: Option<Vec<BucketCors>>

The bucket's Cross-Origin Resource Sharing (CORS) configuration.

owner: Option<BucketOwner>

The owner of the bucket. This is always the project team's owner group.

acl: Option<Vec<BucketAccessControl>>

Access controls on the bucket.

id: Option<String>

The ID of the bucket.

kind: Option<String>

The kind of item this is. For buckets, this is always storage#bucket.

logging: Option<BucketLogging>

The bucket's logging configuration, which defines the destination bucket and optional name prefix for the current bucket's logs.

name: Option<String>

The name of the bucket.

encryption: Option<BucketEncryption>

Encryption configuration used by default for newly inserted objects, when no encryption config is specified.

project_number: Option<String>

The project number of the project the bucket belongs to.

etag: Option<String>

HTTP 1.1 Entity tag for the bucket.

storage_class: Option<String>

The bucket's storage class. This defines how objects in the bucket are stored and determines the SLA and the cost of storage. Values include STANDARD, NEARLINE and DURABLE_REDUCED_AVAILABILITY. Defaults to STANDARD. For more information, see storage classes.

lifecycle: Option<BucketLifecycle>

The bucket's lifecycle configuration. See lifecycle management for more information.

self_link: Option<String>

The URI of this bucket.

location: Option<String>

The location of the bucket. Object data for objects in the bucket resides in physical storage within this region. Defaults to US. See the developer's guide for the authoritative list.