Struct google_storage1::Object

pub struct Object {
    pub self_link: Option<String>,
    pub updated: Option<String>,
    pub content_type: Option<String>,
    pub time_created: Option<String>,
    pub generation: Option<String>,
    pub customer_encryption: Option<ObjectCustomerEncryption>,
    pub component_count: Option<i32>,
    pub metageneration: Option<String>,
    pub media_link: Option<String>,
    pub kms_key_name: Option<String>,
    pub owner: Option<ObjectOwner>,
    pub cache_control: Option<String>,
    pub acl: Option<Vec<ObjectAccessControl>>,
    pub id: Option<String>,
    pub size: Option<String>,
    pub time_deleted: Option<String>,
    pub kind: Option<String>,
    pub content_language: Option<String>,
    pub name: Option<String>,
    pub md5_hash: Option<String>,
    pub bucket: Option<String>,
    pub content_encoding: Option<String>,
    pub crc32c: Option<String>,
    pub etag: Option<String>,
    pub storage_class: Option<String>,
    pub content_disposition: Option<String>,
    pub metadata: Option<HashMap<String, String>>,
}

An object.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list objects (none)

rewrite objects (request)

copy objects (request|response)

watch all objects (none)

get objects (response)

insert objects (request|response)

compose objects (response)

update objects (request|response)

delete objects (none)

patch objects (request|response)

Fields
self_link: Option<String>

The link to this object.

updated: Option<String>

The modification time of the object metadata in RFC 3339 format.

content_type: Option<String>

Content-Type of the object data. If contentType is not specified, object downloads will be served as application/octet-stream.

time_created: Option<String>

The creation time of the object in RFC 3339 format.

generation: Option<String>

The content generation of this object. Used for object versioning.

customer_encryption: Option<ObjectCustomerEncryption>

Metadata of customer-supplied encryption key, if the object is encrypted by such a key.

component_count: Option<i32>

Number of underlying components that make up this object. Components are accumulated by compose operations.

metageneration: Option<String>

The version of the metadata for this object at this generation. Used for preconditions and for detecting changes in metadata. A metageneration number is only meaningful in the context of a particular generation of a particular object.

media_link: Option<String>

Media download link.

kms_key_name: Option<String>

Cloud KMS Key used to encrypt this object, if the object is encrypted by such a key.

owner: Option<ObjectOwner>

The owner of the object. This will always be the uploader of the object.

cache_control: Option<String>

Cache-Control directive for the object data. If omitted, and the object is accessible to all anonymous users, the default will be public, max-age=3600.

acl: Option<Vec<ObjectAccessControl>>

Access controls on the object.

id: Option<String>

The ID of the object.

size: Option<String>

Content-Length of the data in bytes.

time_deleted: Option<String>

The deletion time of the object in RFC 3339 format. Will be returned if and only if this version of the object has been deleted.

kind: Option<String>

The kind of item this is. For objects, this is always storage#object.

content_language: Option<String>

Content-Language of the object data.

name: Option<String>

The name of this object. Required if not specified by URL parameter.

md5_hash: Option<String>

MD5 hash of the data; encoded using base64. For more information about using the MD5 hash, see Hashes and ETags: Best Practices.

bucket: Option<String>

The name of the bucket containing this object.

content_encoding: Option<String>

Content-Encoding of the object data.

crc32c: Option<String>

CRC32c checksum, as described in RFC 4960, Appendix B; encoded using base64 in big-endian byte order. For more information about using the CRC32c checksum, see Hashes and ETags: Best Practices.

etag: Option<String>

HTTP 1.1 Entity tag for the object.

storage_class: Option<String>

Storage class of the object.

content_disposition: Option<String>

Content-Disposition of the object data.

metadata: Option<HashMap<String, String>>

User-provided metadata, in key/value pairs.