Struct google_storage1::ObjectAccessControl

pub struct ObjectAccessControl {
    pub domain: Option<String>,
    pub generation: Option<String>,
    pub object: Option<String>,
    pub bucket: Option<String>,
    pub kind: Option<String>,
    pub entity: Option<String>,
    pub email: Option<String>,
    pub etag: Option<String>,
    pub role: Option<String>,
    pub entity_id: Option<String>,
    pub project_team: Option<ObjectAccessControlProjectTeam>,
    pub id: Option<String>,
    pub self_link: Option<String>,
}

An access-control entry.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get object access controls (response)

patch default object access controls (request|response)

update object access controls (request|response)

list object access controls (none)

patch object access controls (request|response)

update default object access controls (request|response)

insert default object access controls (request|response)

insert object access controls (request|response)

delete object access controls (none)

get default object access controls (response)

Fields

domain

The domain associated with the entity, if any.

generation

The content generation of the object.

object

The name of the object.

bucket

The name of the bucket.

kind

The kind of item this is. For object access control entries, this is always storage#objectAccessControl.

entity

The entity holding the permission, in one of the following forms: - user-userId - user-email - group-groupId - group-email - domain-domain - project-team-projectId - allUsers - allAuthenticatedUsers Examples: - The user liz@example.com would be user-liz@example.com. - The group example@googlegroups.com would be group-example@googlegroups.com. - To refer to all members of the Google Apps for Business domain example.com, the entity would be domain-example.com.

email

The email address associated with the entity, if any.

etag

HTTP 1.1 Entity tag for the access-control entry.

role

The access permission for the entity. Can be READER or OWNER.

entity_id

The ID for the entity, if any.

project_team

The project team associated with the entity, if any.

id

The ID of the access-control entry.

self_link

The link to this access-control entry.